[MSVC] Fix conversion warnings in test.
Fix vector saturate operations.
Fix C-style casts.

+++ test/api/APIVMCoreTest.cpp
@@ -199,7 +199,7 @@ TEST(AsyncRunWsmFile, InterruptTest) {
   WasmEdge_VMContext *VM = WasmEdge_VMCreate(nullptr, nullptr);
   {
     WasmEdge_Async *AsyncCxt = WasmEdge_VMAsyncRunWasmFromBuffer(
-        VM, AsyncWasm.data(), AsyncWasm.size(),
+        VM, AsyncWasm.data(), static_cast<uint32_t>(AsyncWasm.size()),
         WasmEdge_StringWrap("_start", 6), nullptr, 0);
     EXPECT_NE(AsyncCxt, nullptr);
     EXPECT_FALSE(WasmEdge_AsyncWaitFor(AsyncCxt, 1));
@@ -215,7 +215,7 @@ TEST(AsyncRunWsmFile, InterruptTest) {
 TEST(AsyncExecute, InterruptTest) {
   WasmEdge_VMContext *VM = WasmEdge_VMCreate(nullptr, nullptr);
   ASSERT_TRUE(WasmEdge_ResultOK(
-      WasmEdge_VMLoadWasmFromBuffer(VM, AsyncWasm.data(), AsyncWasm.size())));
+      WasmEdge_VMLoadWasmFromBuffer(VM, AsyncWasm.data(), static_cast<uint32_t>(AsyncWasm.size()))));
   ASSERT_TRUE(WasmEdge_ResultOK(WasmEdge_VMValidate(VM)));
   ASSERT_TRUE(WasmEdge_ResultOK(WasmEdge_VMInstantiate(VM)));
   {
